# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: VC Backend API
  version: 1.1.0
  description: |
    Mediasoup-based video-call backend. Covers session lifecycle, join-token
    issuance, ICE servers, customer invite links, recording start/stop,
    session disposition, and admin read-model endpoints.

    All `/v1/*` mutating routes (and most read routes) require an API key
    when `VC_API_KEY` is configured on the server. Send the key as either
    `Authorization: Bearer <key>` or `X-API-Key: <key>`.
servers:
  - url: http://localhost:9000
    description: Local development
...
paths:
...
  /v1/sessions:
    post:
      tags: [sessions]
      summary: Create a new session
      description: |
        Concurrent creates for the same room key (from `metadata.roomName` or
        `externalRef`, trimmed and case-insensitive) return the existing session
        with HTTP 200 and `deduplicated: true`. Stored `metadata.roomName` and
        `externalRef` are normalized to lowercase.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/sessions/resolve:
    get:
      tags: [sessions]
      summary: Resolve a session id by stable room name
      description: |
        Looks up an existing session whose `metadata.roomName` or `externalRef`
        matches the supplied query argument (trimmed, case-insensitive). Returns
        a minimal session header. New sessions store room keys in lowercase.
      security: []
      parameters:
        - in: query
          name: roomName
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Session header
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, status, createdAt, expiresAt]
                properties:
                  sessionId: { type: string }
                  status: { $ref: "#/components/schemas/SessionStatus" }
                  createdAt: { type: string, format: date-time }
                  expiresAt: { type: string, format: date-time }
                  roomName: { type: string, nullable: true }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/sessions/{sessionId}:
    get:
      tags: [sessions]
      summary: Get session state
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/sessions/{sessionId}/join-token:
    post:
      tags: [tokens]
      summary: Issue a participant join token
      description: |
        Returns a short-lived JWT used to authenticate the WebSocket
        `join` event.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinTokenRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/ice-servers:
    get:
      tags: [ice]
      summary: Get STUN/TURN servers for a session
      security: []
      parameters:
        - in: query
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ICE servers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IceServersResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/sessions/{sessionId}/participants:
    get:
      tags: [sessions]
      summary: List participants in a session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Participants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/sessions/{sessionId}/leave:
    post:
      tags: [sessions]
      summary: End a session
      description: |
        Ends the session when at most one participant remains in the store.
        If multiple participants are still present, returns **409**; remove each
        attendee with `POST /v1/sessions/{sessionId}/participants/{participantId}/leave`
        instead so recording and the call can continue for others.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: More than one participant still in the session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sessions/{sessionId}/participants/{participantId}/leave:
    post:
      tags: [sessions]
      summary: Remove one participant from a live session
      description: |
        Participant-scoped leave. Closes that participant's mediasoup resources,
        releases their recording slot (if recording is active), and keeps the
        session alive for remaining participants.
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - in: path
          name: participantId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Participant removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipantLeaveResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Session or participant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sessions/{sessionId}/customer-invite:
    post:
      tags: [invite]
      summary: Issue a customer invite link
      description: |
        Records an invite link on the session for the given participant.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId]
              properties:
                participantId: { type: string }
                channel: { type: string, default: link }
      responses:
        "200":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerInviteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/sessions/{sessionId}/recording/start:
    post:
      tags: [recording]
      summary: Start a recording for the session
      description: |
        Spawns one FFmpeg process per producing participant that writes a
        single `.webm` file each. `initiatedBy` must correspond to a
        participant with `role=agent` once the session has any
        participants (otherwise 403).
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                initiatedBy: { type: string, default: system }
      responses:
        "200":
          description: Recording started
          content:
            application/json:
              schema:
                type: object
                required: [recording]
                properties:
                  recording:
                    $ref: "#/components/schemas/Recording"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot start (already active, no producers, etc.)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      detail: { type: string, nullable: true }
  /v1/sessions/{sessionId}/recording/stop:
    post:
      tags: [recording]
      summary: Stop the active recording
      description: |
        Gracefully signals each per-participant FFmpeg, uploads the
        per-participant WebMs to S3, writes a `manifest.json`, and
        invokes the post-processing Lambda asynchronously. The HTTP
        response reflects the **uploading/uploaded/failed** state at
        the moment the upload finalize hook returns.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                stoppedBy: { type: string, default: system }
      responses:
        "200":
          description: Recording stopped (state in body)
          content:
            application/json:
              schema:
                type: object
                required: [recording]
                properties:
                  recording:
                    $ref: "#/components/schemas/Recording"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: No recording is active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sessions/{sessionId}/disposition:
    post:
      tags: [disposition]
      summary: Record the outcome of a session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              properties:
                outcome:
                  type: string
                  enum: [resolved, follow_up, dropped]
                notes: { type: string, nullable: true }
                resolvedBy: { type: string, nullable: true }
      responses:
        "200":
          description: Disposition recorded
          content:
            application/json:
              schema:
                type: object
                required: [disposition]
                properties:
                  disposition:
                    $ref: "#/components/schemas/Disposition"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/admin/sessions:
    get:
      tags: [admin]
      summary: List recent sessions from the read model
      description: |
        Backed by the optional Postgres read model. Returns 503 when
        `DATABASE_URL` is not configured.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: object
                required: [sessions, total]
                properties:
                  sessions:
                    type: array
                    items: { type: object, additionalProperties: true }
                  total: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Read model disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/admin/sessions/{sessionId}/events:
    get:
      tags: [admin]
      summary: List events for a session from the read model
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, events, total]
                properties:
                  sessionId: { type: string }
                  events:
                    type: array
                    items: { type: object, additionalProperties: true }
                  total: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Read model disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
...
  parameters:
    SessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    SessionStatus:
      type: string
      enum: [created, active, idle, ended]
    CreateSessionRequest:
      type: object
      properties:
        externalRef: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
        ttlSeconds:
          type: integer
          minimum: 60
          maximum: 86400
          default: 3600
    CreateSessionResponse:
      type: object
      required: [sessionId, status, createdAt, expiresAt]
      properties:
        sessionId: { type: string }
        status: { $ref: "#/components/schemas/SessionStatus" }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
...